Added appcontainer to help requests easier
