+++ src/components/UserProfile.js
@@ -1,9 +1,21 @@
-import React from 'react';
+import React, {Component} from 'react';
 
-const UserProfile = (props) => {
+class UserProfile extends Component {
+    constructor(props){
+        super(props);
+        this.state = {
+            userData: undefined
+        }
+    }
+    componentDidMount(){
+        console.log(this.props)
+        // console.log(`Fetching Profile of User ${this.props.match}`)
+    }
+    render(){
     return(
-        <h1>{props.match.params.userId}</h1>
-    )
+        <h1>{JSON.stringify(this.props.match.params.userid)}</h1>
+        )
+    }
 }
 
 export default UserProfile;
